feature/SA-48 Moves normalizer created
