+++ apps/api/src/auth/services/auth.service.ts
@@ -4,7 +4,7 @@ import * as bcrypt from 'bcrypt';
 
 import { User } from '@prisma/client';
 import { UsersService } from 'src/users/services/users.service';
-import { removePassword } from 'src/utils/user';
+import { removePassword } from 'src/utils';
 
 import { SignUpDTO } from '../dtos/signup.dto';
 import { Payload } from '../models/payload.model';
